+++ src/jsontree.js
@@ -15,6 +15,7 @@
     var // Variables: Constructor Parameters
         _parameter_Document = null,
         _parameter_Window = null,
+        _parameter_Navigator = null,
         _parameter_Math = null,
         _parameter_JSON = null,
 
@@ -140,16 +141,28 @@
      */
 
     function renderControlTitleBar( bindingOptions ) {
-        if ( bindingOptions.showTitle || bindingOptions.showTitleTreeControls ) {
-            var titleBar = createElement( bindingOptions.currentView.element, "div", "title-bar" );
+        if ( bindingOptions.showTitle || bindingOptions.showTitleTreeControls || bindingOptions.showTitleCopyButton ) {
+            var titleBar = createElement( bindingOptions.currentView.element, "div", "title-bar" ),
+                controls = createElement( titleBar, "div", "controls" );
         
             if ( bindingOptions.showTitle ) {
-                createElementWithHTML( titleBar, "div", "title", bindingOptions.titleText );
+                createElementWithHTML( titleBar, "div", "title", bindingOptions.titleText, controls );
+            }
+
+            if ( bindingOptions.showTitleCopyButton ) {
+                var copy = createElementWithHTML( controls, "button", "copy-all", _configuration.copyAllButtonText );
+
+                copy.onclick = function() {
+                    var copyData = _parameter_JSON.stringify( _elements_Data[ bindingOptions.currentView.element.id ].data );
+
+                    _parameter_Navigator.clipboard.writeText( copyData );
+
+                    fireCustomTrigger( bindingOptions.onCopy, copyData );
+                };
             }
 
             if ( bindingOptions.showTitleTreeControls ) {
-                var controls = createElement( titleBar, "div", "controls" ),
-                    openAll = createElementWithHTML( controls, "button", "openAll", _configuration.openAllButtonText ),
+                var openAll = createElementWithHTML( controls, "button", "openAll", _configuration.openAllButtonText ),
                     closeAll = createElementWithHTML( controls, "button", "closeAll", _configuration.closeAllButtonText );
 
                 openAll.onclick = function() {
@@ -437,6 +450,7 @@
         options.ignoreFunctionValues = getDefaultBoolean( options.ignoreFunctionValues, false );
         options.reverseArrayValues = getDefaultBoolean( options.reverseArrayValues, false );
         options.addArrayIndexPadding = getDefaultBoolean( options.addArrayIndexPadding, false );
+        options.showTitleCopyButton = getDefaultBoolean( options.showTitleCopyButton, false );
 
         options = buildAttributeOptionStrings( options );
         options = buildAttributeOptionCustomTriggers( options );
@@ -455,6 +469,7 @@
         options.onRenderComplete = getDefaultFunction( options.onRenderComplete, null );
         options.onValueClick = getDefaultFunction( options.onValueClick, null );
         options.onRefresh = getDefaultFunction( options.onRefresh, null );
+        options.onCopy = getDefaultFunction( options.onCopy, null );
 
         return options;
     }
@@ -834,6 +849,7 @@
         _configuration.arrayText = getDefaultString( _configuration.arrayText, "array" );
         _configuration.closeAllButtonText = getDefaultString( _configuration.closeAllButtonText, "Close All" );
         _configuration.openAllButtonText = getDefaultString( _configuration.openAllButtonText, "Open All" );
+        _configuration.copyAllButtonText = getDefaultString( _configuration.copyAllButtonText, "Copy All" );
     }
 
 
@@ -863,9 +879,10 @@
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
      */
 
-    ( function ( documentObject, windowObject, mathObject, jsonObject ) {
+    ( function ( documentObject, windowObject, navigatorObject, mathObject, jsonObject ) {
         _parameter_Document = documentObject;
         _parameter_Window = windowObject;
+        _parameter_Navigator = navigatorObject;
         _parameter_Math = mathObject;
         _parameter_JSON = jsonObject;
 
@@ -879,5 +896,5 @@
             _parameter_Window.$jsontree = this;
         }
 
-    } ) ( document, window, Math, JSON );
+    } ) ( document, window, navigator, Math, JSON );
 } )();